requirements.txt
# Core ML and Data Science
numpy>=1.22,<2.0
pandas>=2.0.0
scikit-learn>=1.3.0
tensorflow>=2.13.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
redis>=4.6.0
streamlit>=1.25.0

# Data Sources
websocket-client>=1.6.0
requests>=2.31.0
polygon-api-client>=1.12.0

# Interpretability
shap>=0.42.0

# Data Processing
pytz>=2023.3
python-dateutil>=2.8.2

# Configuration
python-dotenv>=1.0.0
pydantic>=2.1.0
pydantic-settings>=2.0.0

# MLOps (install separately due to conflicts)
# mlflow>=2.8.0
# feast>=0.50.0

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0